refactor(near-me): extract the chip's locate flow into a shared helper
shareNearMeLocation now owns the fetch → seeded+labelled message →
typed-place fallback sequence (and the fallback dialog) so the chat
composer's upcoming location button can run the identical flow
mid-conversation. NearMeChip keeps its look and behavior exactly;
its onSend type just names the shared NearMeSend typedef.
