+++ src/packages/flutter-app/lib/widgets/near_me_chip.dart
@@ -1,25 +1,18 @@
 import 'package:flutter/material.dart';
 
 import '../l10n/l10n.dart';
-import '../utils/geolocation_stub.dart'
-    if (dart.library.js_interop) '../utils/geolocation_web.dart';
+import 'near_me_locate.dart';
 
 /// "What's near me?" conversation starter (specs-free sibling of the
-/// suggestion chips): one tap grabs a browser geolocation fix and sends a
-/// seeded plan-chat message asking what's good around the traveler right now.
-///
-/// The seed carries the coordinates in its text — the server's search_nearby
-/// tool reads them from the message stream — while [PlanNotifier.sendMessage]'s
-/// `displayLabel` renders it as a compact "Near my current location" context
-/// chip instead of a coordinate-bearing bubble.
-///
-/// When no fix is available (permission denied, non-web build, timeout), a
-/// small dialog asks the traveler to type a city or neighborhood instead; that
-/// branch sends a natural-language message with no coordinates and no label.
+/// suggestion chips): one tap runs the shared [shareNearMeLocation] flow —
+/// a geolocation fix becomes a seeded, labelled plan-chat message asking
+/// what's good around the traveler right now, and the no-fix fallback asks
+/// for a typed neighborhood instead. The chat composer's location button
+/// runs the same flow mid-conversation; this chip stays the opening's way in.
 class NearMeChip extends StatefulWidget {
   /// Delivers the composed message; hosts either send directly on the plan
   /// notifier (Agent tab) or switch tabs first (Home).
-  final void Function(String text, {String? displayLabel}) onSend;
+  final NearMeSend onSend;
 
   /// Styling knobs so the chip can sit among the white-on-photo hero chips as
   /// well as default Material chips. Null keeps the ambient chip theme.
@@ -50,34 +43,11 @@ class NearMeChip extends StatefulWidget {
 class _NearMeChipState extends State<NearMeChip> {
   bool _locating = false;
 
-  Future<void> _locate() async {
-    setState(() => _locating = true);
-    final result = await getCurrentPosition();
-    if (!mounted) return;
-    setState(() => _locating = false);
-
-    final l10n = context.l10n;
-    if (result.ok) {
-      // Coordinates arrive pre-rounded to 4 decimals (~11 m) from the
-      // geolocation util — the exact position never enters the transcript.
-      widget.onSend(
-        l10n.nearMeSeedMessage(
-          result.latitude!.toStringAsFixed(4),
-          result.longitude!.toStringAsFixed(4),
-          (result.accuracyMeters ?? 0).round().toString(),
-        ),
-        displayLabel: l10n.nearMeSeedLabel,
+  Future<void> _locate() => shareNearMeLocation(
+        context,
+        onSend: widget.onSend,
+        onLocating: (locating) => setState(() => _locating = locating),
       );
-      return;
-    }
-
-    final place = await showDialog<String>(
-      context: context,
-      builder: (_) => const _NearMeLocationDialog(),
-    );
-    if (!mounted || place == null || place.trim().isEmpty) return;
-    widget.onSend(context.l10n.nearMeManualMessage(place.trim()));
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -101,74 +71,6 @@ class _NearMeChipState extends State<NearMeChip> {
       backgroundColor: widget.backgroundColor,
       side: widget.backgroundColor != null ? BorderSide.none : null,
       onPressed: _locating ? null : _locate,
-    );
-  }
-}
-
-/// Fallback location entry when no geolocation fix is available. Pops with the
-/// typed place name, or null on cancel.
-class _NearMeLocationDialog extends StatefulWidget {
-  const _NearMeLocationDialog();
-
-  @override
-  State<_NearMeLocationDialog> createState() => _NearMeLocationDialogState();
-}
-
-class _NearMeLocationDialogState extends State<_NearMeLocationDialog> {
-  final _controller = TextEditingController();
-  bool _hasText = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller.addListener(() {
-      final hasText = _controller.text.trim().isNotEmpty;
-      if (hasText != _hasText) setState(() => _hasText = hasText);
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _submit() {
-    final place = _controller.text.trim();
-    if (place.isEmpty) return;
-    Navigator.of(context).pop(place);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    return AlertDialog(
-      title: Text(l10n.nearMeDialogTitle),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(l10n.nearMeDialogMessage),
-          const SizedBox(height: 16),
-          TextField(
-            controller: _controller,
-            autofocus: true,
-            textInputAction: TextInputAction.go,
-            onSubmitted: (_) => _submit(),
-            decoration: InputDecoration(hintText: l10n.nearMeDialogHint),
-          ),
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(),
-          child: Text(l10n.commonCancel),
-        ),
-        FilledButton(
-          onPressed: _hasText ? _submit : null,
-          child: Text(l10n.nearMeDialogCta),
-        ),
-      ],
     );
   }
 }
